+++ Drivers/STM32U5xx_HAL_Driver/Src/stm32u5xx_hal_ospi.c
@@ -1,3 +1,5 @@
+#pragma GCC push_options
+#pragma GCC optimize ("-O0")
 /**
   ******************************************************************************
   * @file    stm32u5xx_hal_ospi.c
@@ -3000,7 +3002,6 @@ static void OSPI_DMAAbortCplt(DMA_HandleTypeDef *hdma)
   * @param  Tickstart : Tick start value
   * @retval HAL status
   */
-__attribute__((optimize("O0")))
 static HAL_StatusTypeDef OSPI_WaitFlagStateUntilTimeout(OSPI_HandleTypeDef *hospi, uint32_t Flag,
                                                         FlagStatus State, uint32_t Tickstart, uint32_t Timeout)
 {
@@ -3032,7 +3033,6 @@ static HAL_StatusTypeDef OSPI_WaitFlagStateUntilTimeout(OSPI_HandleTypeDef *hosp
   * @param  cmd   : structure that contains the command configuration information
   * @retval HAL status
   */
-__attribute__((optimize("O0")))
 static HAL_StatusTypeDef OSPI_ConfigCmd(OSPI_HandleTypeDef *hospi, OSPI_RegularCmdTypeDef *cmd)
 {
   HAL_StatusTypeDef status = HAL_OK;
@@ -3528,3 +3528,5 @@ HAL_StatusTypeDef HAL_OSPI_DLYB_GetClockPeriod(OSPI_HandleTypeDef *hospi, HAL_OS
   */
 
 #endif /* OCTOSPI || OCTOSPI1 || OCTOSPI2 */
+
+#pragma GCC pop_options