+++ Projects/B-U585I-IOT02A/app_compat/barometer_driver.c
@@ -41,6 +41,10 @@ static int get_pressure(float *out_pressure) {
         LOG(ERROR, "Couldn't read pressure");
         return -1;
     }
+    if (!*out_pressure) {
+        // Disregard 0 reading
+        return -1;
+    }
     // convert hPa to Pa
     *out_pressure = 100.0f * *out_pressure;
     return 0;